requirements.txt
appdirs==1.4.0
astroid==1.4.9
click==6.7
coverage==4.3.4
coveralls==1.1
Django~=3.0.7
djangorestframework~=3.11.0
djangorestframework-gis==0.15
docopt==0.6.2
django-allauth==0.42.0
django-environ==0.4.5
django-extensions==2.2.9
dj_rest_auth==1.0.7
factory-boy==2.12.0
Faker==4.1.1
flake8==3.7.9
isort==4.2.5
lazy-object-proxy==1.2.2
mccabe==0.6.1
nose==1.3.7
packaging==16.8
pyparsing==2.1.10
psycopg2-binary~=2.8.5
requests==2.20.0
six==1.10.0
wrapt==1.10.8
